requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-choices==1.6.0
django-environ==0.4.4
django-redis==4.9.0
django-storages==1.6.5
djangorestframework==3.7.7
idna==2.6
oauthlib==2.0.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
python3-openid==3.1.0
pytz==2018.3
raven==6.6.0
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
djangorestframework-gis >= 0.12
Pillow
